requirements.txt
tweepy==4.14.0
requests==2.32.0
python-dotenv==1.0.0
pytest==7.4.0
pytest-asyncio==0.23.5
aiohttp==3.9.3
aiosqlite==0.21.0
flask==3.0.2
flask-wtf==1.2.1
werkzeug==3.0.2
gunicorn==22.0.0
pyjwt==2.8.0
bcrypt==4.1.2
chart.js==4.4.1
psycopg2-binary==2.9.9 